requirements.txt
# Purpose: Lists all Python library dependencies required for the LangGraph agent system.
# Includes core frameworks (FastAPI, LangGraph), LLM integration (LangChain), and service clients.
# Used by pip to install necessary packages during environment setup and container building.

fastapi>=0.110.0
uvicorn>=0.27.0
pydantic-settings>=2.2.0
langchain>=0.1.0
langchain-groq
langchain-community
python-multipart>=0.0.9
httpx>=0.27.0
wikipedia
googlesearch-python
psycopg2-binary
langgraph
python-dotenv
